+++ html.js
@@ -1,56 +1,30 @@
-var isArray = Array.isArray
+var Jsx = require("./lib/jsx")
+var Markup = Jsx.Markup
+var renderChild = Jsx.renderChild
+var renderAttrs = Jsx.renderAttributes
+var typeOf = Jsx.typeOf
 var VOID_ELEMENTS = require("void-elements")
 var VOID_ERR = "Children given to self-closing element: "
 var EMPTY_ARR = Array.prototype
 var DOCTYPE = "<!DOCTYPE html>\n"
-exports = module.exports = jsx
-exports.Fragment = Fragment
+exports = module.exports = Jsx.bind(null, render)
+exports.Fragment = Jsx.Fragment
 exports.Html = Html
 exports.html = newHtml
 
-function jsx(tagName, attrs, children) {
-	switch (typeof tagName) {
-		case "function": return tagName(attrs, children)
-		case "string": return render(tagName, attrs, children)
-		default: throw new TypeError("Tag must be a function or string: " + tagName)
-	}
-}
-
 function render(tagName, attrs, children) {
-	var html = "<" + tagName
-	if (attrs && !isEmpty(attrs)) html += " " + renderAttrs(attrs)
+	var tag = "<" + tagName
+	if (attrs && !isEmpty(attrs)) tag += " " + renderAttrs(attrs)
 
 	if (tagName in VOID_ELEMENTS) {
 		if (children && children.length) throw new RangeError(VOID_ERR + tagName)
-		html += " />"
+		return new Html(tag + " />")
 	}
 	else {
 		var renderTagChild = tagName == "script" ? renderScriptChild : renderChild
 		children = (children || EMPTY_ARR).reduce(renderTagChild, [])
-		html += ">" + children.join("") + "</" + tagName + ">"
+		return new Html(tag + ">" + children.join("") + "</" + tagName + ">")
 	}
-
-	return new Html(html)
-}
-
-function renderChild(children, child) {
-	switch (typeOf(child)) {
-		case "undefined":
-		case "null": break
-		case "boolean":
-		case "number": children.push(child); break
-		case "string": children.push(escapeHtml(child)); break
-		case "array": return child.reduce(renderChild, children)
-
-		case "object":
-			if (child instanceof Html) children.push(String(child))
-			else children.push(escapeHtml(String(child)))
-			break
-
-		default: throw new TypeError("Invalid element: " + child)
-	}
-
-	return children
 }
 
 function renderScriptChild(children, child) {
@@ -77,29 +51,6 @@ function renderScriptChild(children, child) {
 	return children
 }
 
-function renderAttrs(attrs) {
-	var html = [], value
-
-	for (var name in attrs) switch (typeOf(value = attrs[name])) {
-		case "undefined": break;
-		case "null": break;
-		case "boolean": if (value) html.push(name); break
-		case "number": html.push(name + "=\"" + value + "\""); break
-		case "object": value = String(value) // Fall through.
-		case "string": html.push(name + "=\"" + escapeAttr(value) + "\""); break
-		default: throw new TypeError("Invalid attribute value: " + value)
-	}
-
-	return html.join(" ")
-}
-
-function escapeHtml(text) {
-	text = text.replace(/&/g, "&amp;")
-	text = text.replace(/</g, "&lt;")
-	text = text.replace(/>/g, "&gt;")
-	return text
-}
-
 function escapeScript(text) {
 	// https://www.w3.org/TR/html52/semantics-scripting.html#restrictions-for-contents-of-script-elements
 	text = text.replace(/<\/(script)/gi, "<\\/$1")
@@ -107,23 +58,12 @@ function escapeScript(text) {
 	return text
 }
 
-function typeOf(value) {
-	return value === null ? "null" : isArray(value) ? "array" : typeof value
-}
-
-function Fragment(_attrs, children) { return children }
-
 function Html(html) { this.value = html }
-Html.prototype.valueOf = function() { return this.value }
 
-Html.prototype.toString = function(fmt) {
-	switch (fmt) {
-		case undefined: return this.value
-		case "doctype": return DOCTYPE + this.value
-		default: throw new RangeError("Invalid HTML format: " + fmt)
-	}
-}
+Html.prototype = Object.create(Markup.prototype, {
+	constructor: {value: Html, configurable: true, writeable: true},
+	doctype: {value: DOCTYPE, configurable: true, writeable: true}
+})
 
-function escapeAttr(attr) { return attr.replace(/"/g, "&quot;") }
 function isEmpty(obj) { for (var _key in obj) return false; return true }
 function newHtml(html) { return new Html(html) }